MIPS: Allow compressed images to be loaded at any address
Compressed images (vmlinuz.bin) have to be loaded at a specific
address that differ from the address normaly used for vmlinux.bin.
This is because the decompressor just write its output at the address
vmlinux.bin should be loaded at, and it shouldn't overwrite itself.
This limitation mean that the bootloader must be configured differently
when loading a vmlinux.bin or a vmlinuz.bin image, this is annoying
and a source of error.

To workaround this we extend the compressed loader to cope with being
loaded at (nearly) any address. During the early init a jump is used
to compute the offset between the current address and the linked
address, if they differ the whole image is first copied to the linked
address before proceeding.

Some load address won't work, for example if there is an overlap with
the range where vmlinuz.bin should be loaded. However for the typical
case of using the vmlinux.bin address that won't be the case.

Signed-off-by: Alban Bedel <albeu@free.fr>
Suggested-by: Jonas Gorski <jonas.gorski@gmail.com>
---
Changelog:
v2: * Rework the code as suggested by Jonas Gorski to autodetect the
      load address and remove the need for a Kconfig option.